+++ src/components/Button/index.ts
@@ -1,0 +1,4 @@
+export { Button } from "@/components/Button/Button";
+export type { TButtonProps, TButtonVariantProps } from "@/components/Button/Button";
+export { buttonVariants } from "@/components/Button/variants";
+export type { TButtonSize, TButtonVariant } from "@/components/Button/variants";