+++ src/index.ts
@@ -1,0 +1,9 @@
+export { Button, buttonVariants } from "@/components/Button";
+export type {
+    TButtonProps,
+    TButtonSize,
+    TButtonVariant,
+    TButtonVariantProps,
+} from "@/components/Button";
+export { cn } from "@/lib/cn";
+export type { TDistributiveOmit } from "@/lib/types";